Remove dependencies we don't have
The removed dependencies concern GUI support (X11, OpenGL, Wayland, XCB) and
are not needed in headless mode. We don't offer the X11 libraries, and
therefore X11 support, in our main channel anyway. A test testing the GUI is
removed, too.

This is binary repackaging, so the omissions don't affect compilation. There's
also a basic smoke test in the recipe.